+++ contracts/dao/pre-propose/cwd-pre-propose-single-overrule/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod contract;
 mod error;
 mod state;
+mod utils;
 
 #[cfg(test)]
 pub mod testing;
